All vm code moved to X86Sharp.
