--- AssemblySharp/Machine/InstructionType.cs
@@ -1,8 +1,0 @@
-﻿namespace AssemblySharp.Machine
-{
-    public enum InstructionType
-    {
-        Nop,
-        Mov,
-    }
-}
