--- AssemblySharp/Machine/Instructions/DecimalArithmetic.cs
@@ -1,6 +1,0 @@
-﻿namespace AssemblySharp.Machine
-{
-    public partial class VM
-    {
-    }
-}
