--- AssemblySharp/Machine/Instructions/Misc.cs
@@ -1,11 +1,0 @@
-﻿namespace AssemblySharp.Machine
-{
-    public partial class VM
-    {
-        [Instruction(InstructionType.Nop, 0x90)]
-        public void Nop()
-        {
-
-        }
-    }
-}
