--- AssemblySharp/Machine/Instructions/SegmentRegister.cs
@@ -1,6 +1,0 @@
-﻿namespace AssemblySharp.Machine
-{
-    public partial class VM
-    {
-    }
-}
