+++ AssemblySharp/packages.config
@@ -2,5 +2,5 @@
 <packages>
   <package id="System.Memory" version="4.4.0-preview1-25305-02" targetFramework="net47" />
   <package id="System.Numerics.Vectors" version="4.4.0-preview1-25305-02" targetFramework="net47" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.4.0-preview1-25305-02" targetFramework="net47" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="4.4.0" targetFramework="net47" />
 </packages>
