+++ AssemblySharpTests/AssemblySharpTests.csproj
@@ -58,7 +58,6 @@
   </Choose>
   <ItemGroup>
     <Compile Include="InstructionPatternTests.cs" />
-    <Compile Include="Machine\VMTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="X86AssemblyTests.cs" />
   </ItemGroup>
